Tests/LibWeb/Text/expected/css/abspos-inline-static-position-in-pre.txt

one: 0,0
two: 0,20
three: 0,40
right-one: 50,0
